+++ src/rr_types.h
@@ -95,7 +95,7 @@ struct rr_infos {
 extern struct rr_infos rr_info[];
 
 int rr_tp_byname(char *name);
-const char *loc2str(const void *binary, char *ascii, int asclen);
+const char *loc2str(const void *binary, char *ascii, size_t asclen);
 
 #endif
